MFW-1178: Convert to uint before marshalling into protobuf

+++ services/restd/sessions.go
@@ -109,33 +109,33 @@ func parseConntrack(ct *dispatch.Conntrack) map[string]interface{} {
 		return nil
 	}
 
-	m["conntrack_id"] = ct.ConntrackID
-	m["session_id"] = ct.SessionID
-	m["family"] = ct.Family
-	m["ip_protocol"] = ct.ClientSideTuple.Protocol
+	m["conntrack_id"] = uint(ct.ConntrackID)
+	m["session_id"] = uint(ct.SessionID)
+	m["family"] = uint(ct.Family)
+	m["ip_protocol"] = uint(ct.ClientSideTuple.Protocol)
 
-	m["timeout_seconds"] = ct.TimeoutSeconds
-	m["tcp_state"] = ct.TCPState
+	m["timeout_seconds"] = uint(ct.TimeoutSeconds)
+	m["tcp_state"] = uint(ct.TCPState)
 
-	m["client_address"] = ct.ClientSideTuple.ClientAddress
-	m["client_port"] = ct.ClientSideTuple.ClientPort
-	m["server_address"] = ct.ClientSideTuple.ServerAddress
-	m["server_port"] = ct.ClientSideTuple.ServerPort
-	m["client_address_new"] = ct.ServerSideTuple.ClientAddress
-	m["client_port_new"] = ct.ServerSideTuple.ClientPort
-	m["server_address_new"] = ct.ServerSideTuple.ServerAddress
-	m["server_port_new"] = ct.ServerSideTuple.ServerPort
+	m["client_address"] = uint(ct.ClientSideTuple.ClientAddress)
+	m["client_port"] = uint(ct.ClientSideTuple.ClientPort)
+	m["server_address"] = uint(ct.ClientSideTuple.ServerAddress)
+	m["server_port"] = uint(ct.ClientSideTuple.ServerPort)
+	m["client_address_new"] = uint(ct.ServerSideTuple.ClientAddress)
+	m["client_port_new"] = uint(ct.ServerSideTuple.ClientPort)
+	m["server_address_new"] = uint(ct.ServerSideTuple.ServerAddress)
+	m["server_port_new"] = uint(ct.ServerSideTuple.ServerPort)
 
-	m["bytes"] = ct.TotalBytes
-	m["client_bytes"] = ct.ClientBytes
-	m["server_bytes"] = ct.ServerBytes
-	m["packets"] = ct.TotalPackets
-	m["client_packets"] = ct.ClientPackets
-	m["server_packets"] = ct.ServerPackets
+	m["bytes"] = uint(ct.TotalBytes)
+	m["client_bytes"] = uint(ct.ClientBytes)
+	m["server_bytes"] = uint(ct.ServerBytes)
+	m["packets"] = uint(ct.TotalPackets)
+	m["client_packets"] = uint(ct.ClientPackets)
+	m["server_packets"] = uint(ct.ServerPackets)
 
-	m["timestamp_start"] = ct.TimestampStart
+	m["timestamp_start"] = uint(ct.TimestampStart)
 	if ct.TimestampStart != 0 {
-		m["age_milliseconds"] = (uint64(time.Now().UnixNano()) - ct.TimestampStart) / 1000000
+		m["age_milliseconds"] = uint((uint64(time.Now().UnixNano()) - ct.TimestampStart) / 1000000)
 	}
 
 	var mark uint32
@@ -145,12 +145,12 @@ func parseConntrack(ct *dispatch.Conntrack) map[string]interface{} {
 	serverInterfaceID := mark & 0x0000ff00 >> 8
 	serverInterfaceType := mark & 0x0c000000 >> 26
 	priority := mark & 0x00ff0000 >> 16
-	m["mark"] = mark
-	m["client_interface_id"] = clientInterfaceID
-	m["client_interface_type"] = clientInterfaceType
-	m["server_interface_id"] = serverInterfaceID
-	m["server_interface_type"] = serverInterfaceType
-	m["priority"] = priority
+	m["mark"] = uint(mark)
+	m["client_interface_id"] = uint(clientInterfaceID)
+	m["client_interface_type"] = uint(clientInterfaceType)
+	m["server_interface_id"] = uint(serverInterfaceID)
+	m["server_interface_type"] = uint(serverInterfaceType)
+	m["priority"] = uint(priority)
 
 	return m
 }
